share folder with groups
