+++ src/lib/components/dashboard/apis.ts
@@ -1,5 +1,5 @@
 
 export { fetchCredentailsByFolder, fetchCredentialById, addCredential, fetchEncryptedCredentialsFields, shareCredentialsWithUsers, fetchEncryptedFieldsByIds } from '../../apis/credentials.api';
-export { fetchAllFolders, fetchFolderUsers, createFolder, shareFolderWithUsers } from '../../apis/folder.api';
+export { fetchAllFolders, fetchFolderUsers, createFolder, shareFolderWithUsers, shareFolderWithGroups } from '../../apis/folder.api';
 export { fetchAllUserGroups, fetchGroupUsers, createGroup, addUserToGroup, fetchUsersByGroupIds, shareCredentialsWithGroups } from '../../apis/group.api';
 export { fetchAllUsers, createUser } from '../../apis/user.api';
